[resolver] Move repo resolution methods to RepoResolver
